+++ CHANGES.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.2
+- Files 'importUMD.ts' and 'DynamicModule.ts' to load external modules by CDN dynamically. Now 'ScriptLinkLoad.ts' is unused and removed.
+- [Plotly.js](https://plotly.com/javascript/) to generate graphics. "@types/plotly.js" devDependencies installed.
+
 ## 1.2.1
 - Full responsiveness, including light/dark mode.
 - Button and command 'openfile' to open file from computer.